Add tiled-vs-untiled network e2e; fix nlyr guard the live run caught (#38)
Phase 4. Ran the opt-in network e2e against the Planetary Computer — it caught
two things the offline oracles structurally cannot (they never invoke the
network-bound dft_stac_cube):

1. terra::nlyr() returns a double, so the uniform-nlyr guard's vapply template
   integer(1) errored. Fixed to numeric(1). The tiled fetch itself worked — all
   12 tiles streamed, offset-split, and covered without error; only the strict
   guard tripped, after assembly.

2. The tiled and untiled cubes are not co-lattice: gdalcubes enlarges the untiled
   bbox extent symmetrically to align with dx/dy (~0.5px), while tiles anchor at
   the bbox lower-left, so they cannot be compared pixel-for-pixel. Confirmed
   offline on saved real cubes that this is a benign grid offset, not a bug:
   after bilinear alignment cor = 0.997, median |diff| = 3.4e-3, per-layer means
   agree to 6e-4, and mean |diff| at tile seams == interior (no seam artifacts —
   gdalcubes reads the source margin at tile edges). The tiled cube is a faithful
   resampling of the source; downstream reducers are per-pixel, so the offset is
   immaterial.

The network test is rewritten accordingly: a growing-season window for robust
valid-pixel coverage, and equivalence asserted via bilinear alignment
(cor > 0.98, median |diff| < 0.01) plus grid-independent per-layer means
(< 0.01) — thresholds measured on the real cubes with headroom — instead of the
flawed pixel-identity comparison. Offset-split-under-tiling stays covered by the
offline commutativity oracle. Tile anchoring stays at bbox-LL (re-anchoring to
the gdalcubes-enlarged origin would couple to gdalcubes internals and change the
shared #36 helper for no scientific gain).

Relates to NewGraphEnvironment/sred-2025-2026#16

+++ R/dft_stac_cube.R
@@ -335,7 +335,8 @@ dft_stac_cube <- function(aoi,
     tile_stacks <- lapply(tiles, assemble_index_stack)
     # every tile shares t0/t1/dt so nlyr is uniform; guard so a future per-tile
     # time bound fails legibly here rather than deep inside terra::merge.
-    stopifnot(length(unique(vapply(tile_stacks, terra::nlyr, integer(1)))) == 1L)
+    # terra::nlyr() returns a double, so the vapply template is numeric(1).
+    stopifnot(length(unique(vapply(tile_stacks, terra::nlyr, numeric(1)))) == 1L)
     stk <- mosaic_stacks(tile_stacks)
   }
